+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/products/block-editor/create-variable-product-block-editor.spec.js
@@ -12,7 +12,7 @@ const attributes = require( './fixtures/attributes' );
 const tabs = require( './data/tabs' );
 const {
 	waitForGlobalAttributesLoaded,
-} = require( './helpers/confirm-global-attributes-loaded' );
+} = require( './helpers/wait-for-global-attributes-loaded' );
 
 const {
 	createVariableProduct,
