+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/products/block-editor/helpers/wait-for-global-attributes-loaded.js
